# PaperLess.Frontend/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Mock Server",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/"
  } ],
  "paths" : {
...
    "/api/token" : {
      "post" : {
        "operationId" : "GetToken",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            },
            "text/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            },
            "application/*+json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserInfo"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Login" ]
      }
    },
    "/api/search/autocomplete" : {
      "get" : {
        "operationId" : "AutoComplete",
        "parameters" : [ {
          "explode" : true,
          "in" : "query",
          "name" : "term",
          "required" : false,
          "schema" : {
            "type" : "string"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "description" : "Success"
          }
        },
        "tags" : [ "Search" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "UserInfo" : {
        "additionalProperties" : false,
        "example" : {
          "password" : "password",
          "username" : "username"
        },
        "properties" : {
          "username" : {
            "nullable" : true,
            "type" : "string"
          },
          "password" : {
            "nullable" : true,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  }
}